requirements.txt
ipython
jupyterlab
loguru
matplotlib
mkdocs
notebook
numpy
pandas
pip
python-dotenv
ruff
scikit-learn
tqdm
typer
seaborn
openpyxl
plotly
xgboost
statsmodels
scikit-posthocs
optuna
optuna[xgboost]
-e .
